requirements.txt
aioredis
databases[asyncmy]
ddtrace
fastapi[all]
git+https://github.com/akatsuki-v2/shared-modules
httpx
orjson
pre-commit
python-dotenv
python-multipart
structlog
uvicorn
uvloop
